analytics_service/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.41
